[feat] 테이블 컴포넌트를 공용 유틸 컴포넌트로 만들어서 사용, 페이지 네이션 기능 추가
